pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.1"
description = "Python Safety Agent Project"
authors = ["Eddy <eddy1759@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
requests = "^2.32.4"
typer = "^0.16.0"
pip-audit = "^2.9.0"
google-generativeai = "^0.8.5"
black = "^25.1.0"
ruff = "^0.12.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
coverage = "^7.10.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
agent = "agent.cli:app"

[tool.pytest.ini_options]
testpaths = ["."]
addopts = "-v"